+++ src/pages/Home/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Header } from "../components/Header";
+import { Header } from "../../components/Header/Header";
 
 const Home = () => {
   return (
