+++ src/router/RouterConfigure.jsx
@@ -1,13 +1,13 @@
-import Home from "../pages/Home";
-import Page404 from "../pages/Page404";
+import Home from "../pages/Home/Home";
+import Page404 from "../pages/Page404/Page404";
 import SignIn from "../pages/SignIn/SignIn";
 
 const rutas = [
-  // {
-  //   path: "/",
-  //   exact: true,
-  //   page: Home,
-  // },
+  {
+    path: "/home",
+    exact: true,
+    page: Home,
+  },
   {
     path: "/",
     exact: true,
